requirements.txt
numpy==1.23.5
matplotlib==3.7.0
streamlit==1.24.1
Pillow==9.4.0
tensorflow==2.13.0